Reduce props.
